requirements.txt
altgraph==0.17.2
autopep8==2.0.2
beautifulsoup4==4.11.1
bidict==0.22.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cycler==0.11.0
DateTime==5.0
docopt==0.6.2
Flask==2.2.2
Flask-SocketIO==5.3.2
fonttools==4.34.4
future==0.18.2
idna==3.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
kiwisolver==1.4.3
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mpl-finance==0.10.1
mplfinance==0.12.9b7
multitasking==0.0.11
numpy==1.23.0
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.3
pefile==2022.5.30
Pillow==9.2.0
pipreqs==0.4.11
pycodestyle==2.10.0
pyglet==2.0.3
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
pyqtgraph==0.13.2
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tushare==1.2.85
urllib3==1.26.9
websocket-client==0.57.0
webvtt-py==0.4.6
Werkzeug==2.2.2
yarg==0.1.9
yfinance==0.1.74
you-get==0.4.1527
zipp==3.11.0
zope.interface==5.5.2
